Added 2 files.
The 1st file is a helper file that strips any "plain text" copied from the internet into a "form that is usable". i.e. strips all unecessary chars other than [A..Z]
THe 2nd file determines the actual key, given the cipher text and the key-length. Currently the key-length is anywhere from 2 to 30 (but there is not limitation)